requirements.txt
qtpy5
hidapi
